Split hex module
