+++ src/hex/impls.rs
@@ -1,0 +1,286 @@
+use super::Hex;
+use glam::{IVec2, IVec3, Vec2};
+use std::ops::{
+    Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Rem, RemAssign, Sub, SubAssign,
+};
+
+impl Add<Self> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn add(self, rhs: Self) -> Self::Output {
+        self.const_add(rhs)
+    }
+}
+
+impl Add<i32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn add(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x + rhs,
+            y: self.y + rhs,
+        }
+    }
+}
+
+impl AddAssign for Hex {
+    #[inline]
+    fn add_assign(&mut self, rhs: Self) {
+        *self = *self + rhs;
+    }
+}
+
+impl AddAssign<i32> for Hex {
+    #[inline]
+    fn add_assign(&mut self, rhs: i32) {
+        *self = *self + rhs;
+    }
+}
+
+impl Sub<Self> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn sub(self, rhs: Self) -> Self::Output {
+        self.const_sub(rhs)
+    }
+}
+
+impl Sub<i32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn sub(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x - rhs,
+            y: self.y - rhs,
+        }
+    }
+}
+
+impl SubAssign for Hex {
+    #[inline]
+    fn sub_assign(&mut self, rhs: Self) {
+        *self = *self - rhs;
+    }
+}
+
+impl SubAssign<i32> for Hex {
+    #[inline]
+    fn sub_assign(&mut self, rhs: i32) {
+        *self = *self - rhs;
+    }
+}
+
+impl Mul<Self> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn mul(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x * rhs.x,
+            y: self.y * rhs.y,
+        }
+    }
+}
+
+impl Mul<i32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn mul(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x * rhs,
+            y: self.y * rhs,
+        }
+    }
+}
+
+impl Mul<f32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn mul(self, rhs: f32) -> Self::Output {
+        self.rounded_mul(rhs)
+    }
+}
+
+impl MulAssign for Hex {
+    #[inline]
+    fn mul_assign(&mut self, rhs: Self) {
+        *self = *self * rhs;
+    }
+}
+
+impl MulAssign<i32> for Hex {
+    #[inline]
+    fn mul_assign(&mut self, rhs: i32) {
+        *self = *self * rhs;
+    }
+}
+
+impl MulAssign<f32> for Hex {
+    #[inline]
+    fn mul_assign(&mut self, rhs: f32) {
+        *self = *self * rhs;
+    }
+}
+
+impl Div<Self> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn div(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x / rhs.x,
+            y: self.y / rhs.y,
+        }
+    }
+}
+
+impl Div<i32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn div(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x / rhs,
+            y: self.y / rhs,
+        }
+    }
+}
+
+impl Div<f32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn div(self, rhs: f32) -> Self::Output {
+        self.rounded_div(rhs)
+    }
+}
+
+impl DivAssign for Hex {
+    #[inline]
+    fn div_assign(&mut self, rhs: Self) {
+        *self = *self / rhs;
+    }
+}
+
+impl DivAssign<i32> for Hex {
+    #[inline]
+    fn div_assign(&mut self, rhs: i32) {
+        *self = *self / rhs;
+    }
+}
+
+impl DivAssign<f32> for Hex {
+    #[inline]
+    fn div_assign(&mut self, rhs: f32) {
+        *self = *self / rhs;
+    }
+}
+
+impl Rem<Self> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn rem(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x % rhs.x,
+            y: self.y % rhs.y,
+        }
+    }
+}
+
+impl Rem<i32> for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn rem(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x % rhs,
+            y: self.y % rhs,
+        }
+    }
+}
+
+impl RemAssign for Hex {
+    #[inline]
+    fn rem_assign(&mut self, rhs: Self) {
+        *self = *self % rhs;
+    }
+}
+
+impl RemAssign<i32> for Hex {
+    #[inline]
+    fn rem_assign(&mut self, rhs: i32) {
+        *self = *self % rhs;
+    }
+}
+
+impl Neg for Hex {
+    type Output = Self;
+
+    #[inline]
+    fn neg(self) -> Self::Output {
+        self.const_neg()
+    }
+}
+
+impl From<(i32, i32)> for Hex {
+    #[inline]
+    fn from((x, y): (i32, i32)) -> Self {
+        Self { x, y }
+    }
+}
+
+impl From<[i32; 2]> for Hex {
+    #[inline]
+    fn from([x, y]: [i32; 2]) -> Self {
+        Self { x, y }
+    }
+}
+
+impl From<(f32, f32)> for Hex {
+    #[inline]
+    fn from(v: (f32, f32)) -> Self {
+        Self::round(v)
+    }
+}
+
+impl From<[f32; 2]> for Hex {
+    #[inline]
+    fn from([x, y]: [f32; 2]) -> Self {
+        Self::round((x, y))
+    }
+}
+
+impl From<Hex> for IVec2 {
+    #[inline]
+    fn from(hex: Hex) -> Self {
+        hex.as_ivec2()
+    }
+}
+
+impl From<Vec2> for Hex {
+    #[inline]
+    fn from(value: Vec2) -> Self {
+        Self::round((value.x, value.y))
+    }
+}
+
+impl From<Hex> for IVec3 {
+    #[inline]
+    fn from(hex: Hex) -> Self {
+        hex.as_ivec3()
+    }
+}
+
+impl From<IVec2> for Hex {
+    #[inline]
+    fn from(v: IVec2) -> Self {
+        Self::new(v.x, v.y)
+    }
+}
